--- components/base64/CMakeLists.txt
@@ -1,6 +1,0 @@
-idf_component_register(
-SRCS
-  "base64.cc"
-INCLUDE_DIRS
-  "."
-)
