+++ components/blob/include/blob/base64.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include "bytes.h"
+
+namespace Blob
+{
+  std::string Base64Encode(const Bytes& bytes);
+  Bytes Base64Decode(const std::string& b64);
+}
