+++ components/blob/include/blob/bytes.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include <string>
+
+namespace Blob
+{
+  typedef std::basic_string<uint8_t> Bytes;
+}
