+++ main/CMakeLists.txt
@@ -1,12 +1,7 @@
-set(SRCS
-  main.cc
-)
-
 idf_component_register(
 SRCS
-  "${SRCS}"
-REQUIRES
-  esp_timer heap esp_netif esp_eth esp_wifi t_amoled nvs_flash # assets lvgl
+  "main.cc"
 PRIV_REQUIRES
-  wolfssh wolfssl peripheral
+  esp_timer heap # assets lvgl
+  peripheral ssh
 )
